Remove unused code and move to bpf files when needed
