+++ pkg/security/probe/exec_bpf.go
@@ -1,0 +1,14 @@
+// Unless explicitly stated otherwise all files in this repository are licensed
+// under the Apache License Version 2.0.
+// This product includes software developed at Datadog (https://www.datadoghq.com/).
+// Copyright 2016-2020 Datadog, Inc.
+
+// +build linux_bpf
+
+package probe
+
+// execTables holds the list of eBPF tables used by the process kprobes
+var execTables = []string{
+	"proc_cache",
+	"pid_cookie",
+}
