+++ pkg/security/probe/open.go
@@ -11,16 +11,6 @@ import (
 	"github.com/DataDog/datadog-agent/pkg/security/secl/eval"
 )
 
-// openTables is the list of eBPF tables used by open's kProbes
-var openTables = []string{
-	"open_policy",
-	"open_basename_approvers",
-	"open_flags_approvers",
-	"open_flags_discarders",
-	"open_process_inode_approvers",
-	"open_path_inode_discarders",
-}
-
 var openCapabilities = Capabilities{
 	"open.filename": {
 		PolicyFlags:     PolicyFlagBasename,
